LayoutTests/http/wpt/webrtc/audiovideo-script-transform.html

<!doctype html>
<html>
    <head>
        <meta charset="utf-8">
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
    </head>
    <body>
        <video id="video" autoplay playsInline></video>
        <script src="routines.js"></script>
        <script>
function waitForExpectedMessage(port, expected)
{
    let gotMessage;
    const promise = new Promise((resolve, reject) => {
        gotMessage = resolve;
        setTimeout(() => { reject("did not get '" + data + "' last message data is '" + promise.lastData + "'") }, 5000);
    });
    port.onmessage = event => {
       promise.lastData = event.data;
       if (event.data === expected)
           gotMessage();
    };
    return promise;
}

function waitForMessage(port)
{
    let gotMessage;
    const promise = new Promise((resolve, reject) => {
        gotMessage = resolve;
        setTimeout(() => { reject("did not get any message") }, 5000);
    });
    port.onmessage = event => gotMessage(event.data);
    return promise;
}

function createTransforms(worker, mediaType)
{
    const senderChannel = new MessageChannel;
    const receiverChannel = new MessageChannel;
    senderTransform = new RTCRtpScriptTransform(worker, {mediaType: mediaType, side: 'sender', port: senderChannel.port2}, [senderChannel.port2]);
    receiverTransform = new RTCRtpScriptTransform(worker, {mediaType: mediaType, side: 'receiver', port: receiverChannel.port2}, [receiverChannel.port2]);
    senderTransform.port = senderChannel.port1;
    receiverTransform.port = receiverChannel.port1;
    return [senderTransform, receiverTransform];
}

let localStream, videoSenderTransform, videoReceiverTransform, audioSenderTransform, audioReceiverTransform;
promise_test(async (test) => {
    worker = new Worker('audio-video-transform.js');
    const data = await new Promise(resolve => worker.onmessage = (event) => resolve(event.data));
    assert_equals(data, "registered");

    localStream = await navigator.mediaDevices.getUserMedia({audio: true, video: true});

    [audioSenderTransform, audioReceiverTransform] = createTransforms(worker, 'audio');
    [videoSenderTransform, videoReceiverTransform] = createTransforms(worker, 'video');

    const stream = await new Promise((resolve, reject) => {
        createConnections((firstConnection) => {
            const audioSender = firstConnection.addTrack(localStream.getAudioTracks()[0], localStream);
            audioSender.transform = audioSenderTransform;
            const videoSender = firstConnection.addTrack(localStream.getVideoTracks()[0], localStream);
            videoSender.transform = videoSenderTransform;
        }, (secondConnection) => {
            secondConnection.ontrack = (trackEvent) => {
                receiver = trackEvent.receiver;
                receiver.transform = trackEvent.track.kind === "audio" ? audioReceiverTransform : videoReceiverTransform;
                resolve(trackEvent.streams[0]);
            };
        });
        test.step_timeout(() => reject("Test timed out"), 5000);
    });

    video.srcObject = stream;
    return video.play();
}, "setup");

promise_test((test) => {
    audioSenderTransform.port.postMessage("tryGenerateKeyFrame");
    return waitForExpectedMessage(audioSenderTransform.port, "FAIL: InvalidStateError");
}, "key frame on audio sender");

promise_test((test) => {
    audioReceiverTransform.port.postMessage("trySendKeyFrameRequest");
    return waitForExpectedMessage(audioReceiverTransform.port, "FAIL: InvalidStateError");
}, "key frame on audio receiver");

promise_test((test) => {
    videoReceiverTransform.port.postMessage("tryGenerateKeyFrame");
    return waitForExpectedMessage(videoReceiverTransform.port, "FAIL: InvalidStateError");
}, "generate key frame on video receiver");

promise_test((test) => {
    videoSenderTransform.port.postMessage("trySendKeyFrameRequest");
    return waitForExpectedMessage(videoSenderTransform.port, "FAIL: InvalidStateError");
}, "send request key frame on video sender");

promise_test((test) => {
    videoSenderTransform.port.postMessage("tryGenerateKeyFramePromiseHandling");
    return waitForExpectedMessage(videoSenderTransform.port, "PASS");
}, "generate key frame promise handling on video sender");

promise_test(async (test) => {
    audioReceiverTransform.port.postMessage("checkDataAfterWrite");
    await waitForMessage(audioReceiverTransform.port, "PASS");

    audioSenderTransform.port.postMessage("checkDataAfterWrite");
    await waitForMessage(audioSenderTransform.port, "PASS");

    videoReceiverTransform.port.postMessage("checkDataAfterWrite");
    await waitForMessage(videoReceiverTransform.port, "PASS");

    videoSenderTransform.port.postMessage("checkDataAfterWrite");
    await waitForMessage(videoSenderTransform.port, "PASS");
}, "Check data after write");

promise_test(async (test) => {
    audioReceiverTransform.port.postMessage("checkModifiedDataAfterWrite");
    await waitForMessage(audioReceiverTransform.port, "PASS");

    audioSenderTransform.port.postMessage("checkModifiedDataAfterWrite");
    await waitForMessage(audioSenderTransform.port, "PASS");

    videoReceiverTransform.port.postMessage("checkModifiedDataAfterWrite");
    await waitForMessage(videoReceiverTransform.port, "PASS");

    videoSenderTransform.port.postMessage("checkModifiedDataAfterWrite");
    await waitForMessage(videoSenderTransform.port, "PASS");
}, "Check modified data after write");

promise_test(async (test) => {
    audioReceiverTransform.port.postMessage("tryAccessingDataTwice");
    await waitForExpectedMessage(audioReceiverTransform.port, "PASS");

    audioSenderTransform.port.postMessage("tryAccessingDataTwice");
    await waitForExpectedMessage(audioSenderTransform.port, "PASS");

    videoReceiverTransform.port.postMessage("tryAccessingDataTwice");
    await waitForExpectedMessage(videoReceiverTransform.port, "PASS");

    videoSenderTransform.port.postMessage("tryAccessingDataTwice");
    await waitForExpectedMessage(videoSenderTransform.port, "PASS");
}, "Check data getter");

promise_test(async (test) => {
    audioReceiverTransform.port.postMessage("tryAccessingMetadata");
    let metadata = await waitForMessage(audioReceiverTransform.port);
    assert_array_equals(Object.keys(metadata), ["contributingSources", "mimeType", "payloadType", "rtpTimestamp", "sequenceNumber", "synchronizationSource"]);

    audioSenderTransform.port.postMessage("tryAccessingMetadata");
    metadata = await waitForMessage(audioSenderTransform.port);
    assert_array_equals(Object.keys(metadata), ["contributingSources", "mimeType", "payloadType", "rtpTimestamp", "synchronizationSource"]);

    videoReceiverTransform.port.postMessage("tryAccessingMetadata");
    metadata = await waitForMessage(videoReceiverTransform.port);
    assert_array_equals(Object.keys(metadata), ["contributingSources", "dependencies", "height", "mimeType", "payloadType", "rtpTimestamp", "spatialIndex", "synchronizationSource", "temporalIndex", "width"]);

    videoSenderTransform.port.postMessage("tryAccessingMetadata");
    metadata =await waitForMessage(videoSenderTransform.port);
    assert_array_equals(Object.keys(metadata), ["contributingSources", "dependencies", "frameId", "height", "mimeType", "payloadType", "rtpTimestamp", "spatialIndex", "synchronizationSource", "temporalIndex", "width"]);

    test.add_cleanup(() => localStream.getTracks().forEach(t => t.stop()));
}, "Check metadata getter");
        </script>
    </body>
</html>
